ticket:8822 merge from trunk r30691 through r32118

+++ icu4j/branches/srl/islamic8822/main/classes/core/src/com/ibm/icu/impl/Utility.java
@@ -78,7 +78,7 @@ public final class Utility {
         if (source instanceof int[])
             return(arrayEquals((int[]) source,target));
         if (source instanceof double[])
-            return(arrayEquals((int[]) source,target));
+            return(arrayEquals((double[]) source, target));
         if (source instanceof byte[])
             return(arrayEquals((byte[]) source,target));
         return source.equals(target);
